clang-tidy fixes

+++ tasks/rozenberg_a_matrix_column_sum/mpi/src/ops_mpi.cpp
@@ -2,6 +2,7 @@
 
 #include <mpi.h>
 
+#include <cstddef>
 #include <vector>
 
 #include "rozenberg_a_matrix_column_sum/common/include/common.hpp"
@@ -37,8 +38,8 @@ bool RozenbergAMatrixColumnSumMPI::PreProcessingImpl() {
     size_t rows = GetInput().size();
     size_t columns = GetInput()[0].size();
     flat_.resize(rows * columns);
-    for (int i = 0; i < static_cast<int>(rows); i++) {
-      for (int j = 0; j < static_cast<int>(columns); j++) {
+    for (size_t i = 0; i < static_cast<int>(rows); i++) {
+      for (size_t j = 0; j < static_cast<int>(columns); j++) {
         flat_[j + (i * columns)] = GetInput()[i][j];
       }
     }
